# Explicit loop in `_findall` to avoid allocations