While testing #2724 and looking at #2769 benchmarks I've seen that sometimes `_findall` allocates a lot and has some weird  overhead in memory. After digging in, changing
`I[i:i+64-1] .= i1:(i1+64-1)` 
into 
```
for j in 0:64-1
    I[i + j] = i1 + j 
end
```
removed unnecessary allocations, and sped up execution a bit.
I'll post benchmarks tommorow.