@bkamins 2021-05-23
This is strange. @mbauman - broadcasting in these cases should be fast - right? The only potential benefit should be that we avoid extra checking of dimensions of the broadcasted objects, but no allocations should be made I think.

Still we might want to remove it to reduce compilation latency, but I would like to understand the reason of the problem with broadcasting.