@pstorozenko 2021-05-23
It is weird for me as I checked at some point number of allocations with `I[i:i+64-1] .= i1:(i1+64-1)` and it was zero when working on `_findall`.
My theory is that range `i1:(i1+64-1)` is collected here and then later assigned but I might be totally wrong.