@bkamins 2021-05-23
`I[i:i+64-1] .= i1:(i1+64-1)` should be zero allocations and not collect anything. That is why I have pinged @mbauman.