@pstorozenko 2021-05-25
Thanks, do we know what is the cause of allocations?