@bkamins 2021-05-25
No. It is very strange. If I use Revise.jl and modify the body of the function in an unrelated place then in your old code I go from:
```
DataFrames._findall(B) = true
  219.265 ns (5 allocations: 1.61 KiB)
```
to 1 allocation (as in your code with broadcasting replaced by a loop). So it seems that for some reason the compiler gets confused and is not generating optimal code.

@vtjnash - would you be interested in reporting the details? (for our use case it is a minor issue so merged the PR, but maybe for some reason you would find it useful to have reproduced)